Extract domain input row and add unit tests
