+++ src/web/components/ClientSideTokenGeneration/CstgDomainInputRow.scss
@@ -1,0 +1,28 @@
+.cstg-domains-table {
+  .domain-input-remove-button {
+    padding: 0 0 0 2px;
+
+    svg {
+      color: var(--theme-table-action-button);
+    }
+  }
+
+  .domain-input-field {
+    display: flex;
+    gap: 10px;
+    align-items: center;
+
+    .input-container {
+      border-radius: 5px;
+      border: 1px solid var(--theme-input-border);
+    }
+
+    .input-container.Warning {
+      border: 1px solid var(--theme-warning-border);
+    }
+
+    .input-container.Error {
+      border: 1px solid var(--theme-error-border);
+    }
+  }
+}
